Extract fallback questions repository as own class
